requirements.txt
Flask==2.3.3
transformers==4.34.0
torch==2.3.1
torchvision==0.18.1
Pillow==10.0.0
numpy==1.26.4
requests==2.31.0